UI Transition
Since all of the math for the map editor relies on screen space overlay AND the rest of the UI needs to hover above the map, I now needed to move all of the UI screens except the 3D dice setup into screen space overlay. Kind of a pain, but worth it for the map editor system.