MapItemList & MapItem Implemented
These two classes allow the user to browse assets that can be placed on the map. When the user clicks one a copy of it is added to the map in the first available space from the top left. The user can then pick it up and place it wherever they'd like.